[CBRD-21528] separate partition.h into server's and both's
