+++ src/parser/semantic_check.c
@@ -41,6 +41,7 @@
 #include "xasl_generation.h"
 #include "view_transform.h"
 #include "show_meta.h"
+#include "partition.h"
 
 /* this must be the last header file included!!! */
 #include "dbval.h"
