+++ src/query/execute_schema.h
@@ -46,7 +46,6 @@
 #define PARTITION_VARCHAR_LEN (DB_MAX_IDENTIFIER_LENGTH)
 #define CLASS_ATT_NAME "class_name"
 #define CLASS_IS_PARTITION "partition_of"
-#define MAX_PARTITIONS 1024
 
 #define CHECK_PARTITION_NONE    0x0000
 #define CHECK_PARTITION_PARENT  0x0001
